requirements.txt
vunit_hdl>=4.7.0
colorama>=0.4.0
